take password length setting ui out of the dialog
